+++ src/com/amplitude/api/DeviceInfo.java
@@ -156,6 +156,8 @@ public class DeviceInfo {
                     // failed to fetch geocoder
                 } catch (IllegalArgumentException e) {
                     // Bad lat / lon values can cause Geocoder to throw IllegalArgumentExceptions
+                } catch (IllegalStateException e) {
+                    // sometimes the location manager is unavailable
                 }
             }
             return null;